backend/requirements.txt
flask
requests
geopy
timezonefinder
pytz
flask_sqlalchemy
flask_cors
flask_praetorian
sqlalchemy
flask_apscheduler
flask_mailman
pytest
python-dotenv